pyproject.toml
[project]
name = "whisper-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.123.1",
    "huey>=2.5.4",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "numpy<2.0",
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uuid>=1.30",
    "uvicorn>=0.38.0",
]
